chore(dictionary): cleaning unused code
