+++ utils/constants/categories.js
@@ -1,0 +1,17 @@
+const CATEGORIES = {
+  semua: "semua",
+  angka: "angka",
+  bulan: "bulan",
+  hari: "hari",
+  hewan: "hewan",
+  keluarga: "keluarga",
+  kerja: "kerja",
+  kota: "kota",
+  olahraga: "olahraga",
+  perkenalan: "perkenalan",
+  sifat: "sifat",
+};
+
+Object.freeze(CATEGORIES);
+
+export { CATEGORIES };
