+++ utils/constants/dictionary_items.js
@@ -1,0 +1,34 @@
+import { CATEGORIES } from "./categories";
+
+export const DICT_ITEMS = [
+  { label: "1", category: CATEGORIES.angka },
+  { label: "2", category: CATEGORIES.angka },
+  { label: "3", category: CATEGORIES.angka },
+  { label: "4", category: CATEGORIES.angka },
+  { label: "5", category: CATEGORIES.angka },
+  { label: "6", category: CATEGORIES.angka },
+  { label: "7", category: CATEGORIES.angka },
+  { label: "8", category: CATEGORIES.angka },
+  { label: "9", category: CATEGORIES.angka },
+  { label: "10", category: CATEGORIES.angka },
+  { label: "angka", category: CATEGORIES.angka },
+  { label: "januari", category: CATEGORIES.bulan },
+  { label: "februari", category: CATEGORIES.bulan },
+  { label: "maret", category: CATEGORIES.bulan },
+  { label: "april", category: CATEGORIES.bulan },
+  { label: "mei", category: CATEGORIES.bulan },
+  { label: "juni", category: CATEGORIES.bulan },
+  { label: "juli", category: CATEGORIES.bulan },
+  { label: "agustus", category: CATEGORIES.bulan },
+  { label: "september", category: CATEGORIES.bulan },
+  { label: "oktober", category: CATEGORIES.bulan },
+  { label: "november", category: CATEGORIES.bulan },
+  { label: "desember", category: CATEGORIES.bulan },
+  { label: "senin", category: CATEGORIES.hari },
+  { label: "selasa", category: CATEGORIES.hari },
+  { label: "rabu", category: CATEGORIES.hari },
+  { label: "kamis", category: CATEGORIES.hari },
+  { label: "jumat", category: CATEGORIES.hari },
+  { label: "sabtu", category: CATEGORIES.hari },
+  { label: "minggu", category: CATEGORIES.hari },
+];
