+++ utils/constants/index.js
@@ -1,0 +1,2 @@
+export * from "./categories";
+export * from "./dictionary_items";
